requirements.txt
alembic==1.17.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
beartype==0.22.2
billiard==4.2.2
black==25.9.0
boolean.py==5.0
boto3==1.40.55
botocore==1.40.55
celery==5.5.3
certifi==2025.10.5
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cyclonedx-python-lib==11.2.0
defusedxml==0.7.1
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jmespath==1.0.1
kombu==5.5.4
license-expression==30.4.4
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packageurl-python==0.17.5
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
ply==3.11
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-serializable==2.1.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
PyYAML==6.0.3
rdflib==7.2.1
redis==5.2.1
ruff==0.14.1
s3transfer==0.14.0
semantic-version==2.10.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
spdx-tools==0.8.3
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
xmltodict==1.0.2
aiofiles==25.1.0
python-multipart==0.0.20